+++ src/Beef.Core/BusinessException.cs
@@ -12,6 +12,8 @@ namespace Beef
     /// <para>The <see cref="Exception.Message"/> defaults to: <i>A business error occurred.</i></para></remarks>
     public class BusinessException : Exception, IBusinessException
     {
+        private const string _message = "A business error occurred.";
+
         /// <summary>
         /// Get or sets the <see cref="ShouldBeLogged"/> value.
         /// </summary>
@@ -26,43 +28,28 @@ namespace Beef
         /// Initializes a new instance of the <see cref="BusinessException"/> class with a specified messsage.
         /// </summary>
         /// <param name="message">The message text.</param>
-        public BusinessException(string? message)
-            : base(message ?? new LText("Beef.BusinessException"))
-        {
-        }
+        public BusinessException(string? message) : base(message ?? new LText("Beef.BusinessException", _message)) { }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="BusinessException"/> class with a specified messsage and inner exception.
         /// </summary>
         /// <param name="message">The message text.</param>
         /// <param name="innerException">The inner <see cref="Exception"/>.</param>
-        public BusinessException(string? message, Exception innerException)
-            : base(message ?? new LText("Beef.BusinessException"), innerException)
-        {
-        }
+        public BusinessException(string? message, Exception innerException) : base(message ?? new LText("Beef.BusinessException", _message), innerException) { }
 
         /// <summary>
         /// Gets the <see cref="ErrorType"/> (see <see cref="ErrorType.BusinessError"/>).
         /// </summary>
-        public ErrorType ErrorType
-        {
-            get { return ErrorType.BusinessError; }
-        }
+        public ErrorType ErrorType => ErrorType.BusinessError; 
 
         /// <summary>
         /// Gets the corresponding <see cref="HttpStatusCode"/>.
         /// </summary>
-        public HttpStatusCode StatusCode
-        {
-            get { return HttpStatusCode.BadRequest; }
-        }
+        public HttpStatusCode StatusCode => HttpStatusCode.BadRequest; 
 
         /// <summary>
         /// Indicates whether the <see cref="Exception"/> should be logged (returns the <see cref="ShouldExceptionBeLogged"/> value).
         /// </summary>
-        public bool ShouldBeLogged
-        {
-            get { return ShouldExceptionBeLogged; }
-        }
+        public bool ShouldBeLogged => ShouldExceptionBeLogged; 
     }
 }