+++ src/Beef.Core/ConcurrencyException.cs
@@ -11,6 +11,8 @@ namespace Beef
     /// <remarks>The <see cref="Exception.Message"/> defaults to: <i>A concurrency error occurred; please refresh the data and try again.</i></remarks>
     public class ConcurrencyException : Exception, IBusinessException
     {
+        private const string _message = "A concurrency error occurred; please refresh the data and try again.";
+
         /// <summary>
         /// Get or sets the <see cref="ShouldBeLogged"/> value.
         /// </summary>
@@ -19,52 +21,34 @@ namespace Beef
         /// <summary>
         /// Initializes a new instance of the <see cref="ConcurrencyException"/> class.
         /// </summary>
-        public ConcurrencyException()
-            : this(null!)
-        {
-        }
+        public ConcurrencyException() : this(null!) { }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="ConcurrencyException"/> class with a specified messsage.
         /// </summary>
         /// <param name="message">The message text.</param>
-        public ConcurrencyException(string? message)
-            : base(message ?? new LText("Beef.ConcurrencyException"))
-        {
-        }
+        public ConcurrencyException(string? message) : base(message ?? new LText("Beef.ConcurrencyException", _message)) { }
 
         /// <summary>
         /// Initializes a new instance of the <see cref="ConcurrencyException"/> class with a specified messsage and inner exception.
         /// </summary>
         /// <param name="message">The message text.</param>
         /// <param name="innerException">The inner <see cref="Exception"/>.</param>
-        public ConcurrencyException(string? message, Exception innerException)
-            : base(message ?? new LText("Beef.ConcurrencyException"), innerException)
-        {
-        }
+        public ConcurrencyException(string? message, Exception innerException) : base(message ?? new LText("Beef.ConcurrencyException", _message), innerException) { }
 
         /// <summary>
         /// Gets the <see cref="ErrorType"/> (see <see cref="ErrorType.ConcurrencyError"/>).
         /// </summary>
-        public ErrorType ErrorType
-        {
-            get { return ErrorType.ConcurrencyError; }
-        }
+        public ErrorType ErrorType => ErrorType.ConcurrencyError; 
 
         /// <summary>
         /// Gets the corresponding <see cref="HttpStatusCode"/>.
         /// </summary>
-        public HttpStatusCode StatusCode
-        {
-            get { return HttpStatusCode.PreconditionFailed; }
-        }
+        public HttpStatusCode StatusCode => HttpStatusCode.PreconditionFailed; 
 
         /// <summary>
         /// Indicates whether the <see cref="Exception"/> should be logged (returns the <see cref="ShouldExceptionBeLogged"/> value).
         /// </summary>
-        public bool ShouldBeLogged
-        {
-            get { return ShouldExceptionBeLogged; }
-        }
+        public bool ShouldBeLogged => ShouldExceptionBeLogged; 
     }
 }